move command to ClientProxy add stop command to cross show progress when try too large

+++ src/main/java/city/windmill/cropchance/command/ChatBuilder.java
@@ -162,7 +162,7 @@ public class ChatBuilder {
         return this;
     }
 
-    public ChatBuilder addAttr(String name, int value) {
+    public ChatBuilder addAttr(String name, long value) {
         return addAttr(name, String.format("%d", value));
     }
